requirements.txt
# TrendToThread - Twitter MCP Server
# Python dependencies

mcp[cli]>=1.0.0
requests>=2.31.0
requests-oauthlib>=1.3.1
python-dotenv>=1.0.0
pyngrok>=7.0.0
uvicorn>=0.27.0
starlette>=0.36.0
